type:[#test] Test component 추가
