+++ front/src/RouterList.jsx
@@ -1,17 +1,14 @@
-// import {Layout} from "./layout/Layout.jsx";
+// import { Layout } from "./layout/Layout.jsx";
 import MainPage from "./pages/MainPage.jsx";
-import {ChatPage} from "./pages/ChatPage.jsx"
-// import {Test} from "./pages/Test.jsx";
+import { ChatPage } from "./pages/ChatPage.jsx";
+import { Test } from "./pages/Test.jsx";
 import { createBrowserRouter } from "react-router-dom";
 import { MapPage } from "./pages/MapPage.jsx";
-
-
 
 export const RouterList = () => [
     {
         path: "/",
         element: <MainPage />,
-
     },
     {
         path: "/map",
@@ -21,6 +18,15 @@ export const RouterList = () => [
         path: "/chat",
         element: <ChatPage />,
     },
-]
+    {
+        path: "/login",
+        children: [
+            {
+                path: "done",
+                element: <Test />,
+            },
+        ],
+    },
+];
 
 export const RouterObject = createBrowserRouter(RouterList());
